Add files via upload
This is my working attempt at using the while loop to carve from every item in the SOF list to every EOF list item that it is greater than.  If you run it, you'll see it works for the first, but then I can't get the EOF to reset back to the first item in the list.  It stays as the last.  I'll keep working on it, but wanted to give you what I had so far, so you can mess with this one.  Then we can append to the final copy once this chunk is worked out.